% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_estimates-class.R
\name{estimate_biomass}
\alias{estimate_biomass}
\alias{is.biomass_estimates}
\alias{print.biomass_estimates}
\alias{plot.biomass_estimates}
\title{Estimate biomass for each node in a food web}
\usage{
estimate_biomass(production_estimates, pb_ratio)

is.biomass_estimates(x)

\method{print}{biomass_estimates}(x, ...)

\method{plot}{biomass_estimates}(x, nodes = NULL, settings = list(), ...)
}
\arguments{
\item{production_estimates}{a production_estimates object calculated with \link[trophic]{production_estimates}}

\item{pb_ratio}{a production:biomass ratio object created with \link[trophic]{pb_ratio}}

\item{x}{a biomass_estimates object}

\item{...}{further arguments passed to or from other methods}

\item{nodes}{integer or integer vector of nodes to plot (indexed by food_web row)}

\item{settings}{plot settings passed directly to \link[graphics]{plot}}
}
\value{
An object of class \code{biomass_estimates}
}
\description{
Estimation converts production estimates to biomass estimates
}
\examples{

library(trophic)
library(future)
plan(multiprocess)

# Load pre-compiled food_web, efficiency_matrix, and dominance_matrix objects

food_web <- data(example_food_web)
dominance <- data(example_dominance_matrix)
efficiency_mean <- data(example_efficiency_matrix)
efficiency_sd <- data(example_efficiency_sd)

# Construct the component objects
test_fw <- build_food_web(interaction_matrix = food_web)
test_efficiency_matrix <- build_efficiency_matrix(efficiency_mean = efficiency_mean,
                                                  efficiency_sd = efficiency_sd)
test_dominance <- build_dominance_matrix(dominance = dominance)
test_primary_producers <- build_primary_producers(production_mean = c(1, 2),
                                                  production_sd = c(0.5, 0.5))

# Construct the trophic_dynamics object
test_trophic_dynamics <- build_trophic_dynamics(food_web = test_fw,
                                                efficiency_matrix = test_efficiency_matrix,
                                                dominance_matrix = test_dominance)

# Estimate production values from constructed trophic_dynamics object
production_estimates <- estimate_production(test_trophic_dynamics, test_primary_producers)

# Create a pb_ratio object
test_pb_ratio <- build_pb_ratio(range = c(0.25, 5.75), probs = c(5, 20, 10, 3, 1, 1, 1))

# Convert production to biomass estimates
biomass_estimates <- estimate_biomass(production_estimates, test_pb_ratio)

# Test if object is of the type 'biomass_estimates'
  
is.biomass_estimates(x)

# Print information about the 'biomass_estimates' object

print(x)

# Plot a 'biomass_estimates' object

plot(x)
}
