% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{cascade}
\alias{cascade}
\title{Cascade Food Web Model}
\usage{
cascade(S, C)
}
\arguments{
\item{S}{Number of species in the community.}

\item{C}{Average path length used to get connectance.}
}
\value{
The adjacency matrix of a cascade model food web. Species are arranged along a niche axis and may consume any species with a niche value less than their own with a fixed probability of \code{2CS/(S-1)}, creating a triangular matrix.
}
\description{
Cascade Food Web Model
}
\examples{
cascade(20, .1)
}
