% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{cascade_II}
\alias{cascade_II}
\title{Cascade Food Web Model Type II}
\usage{
cascade_II(S, C)
}
\arguments{
\item{S}{Number of species in the community.}

\item{C}{Average path length used to get connectance.}
}
\value{
The adjacency matrix of a cascade model food web where species higher on the niche axis feed on those lower with a fixed probability \code{C/S}.
}
\description{
Cascade Food Web Model Type II
}
\section{Reference}{

Cohen, J. E., and C. M. Newman. 1985. A stochastic theory of community food webs: I. Models and aggregated data. Proceedings of the Royal Society B 224:421–448.
}

\examples{
cascade_II(20, 1.86)
}
