% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{cascade_likelihood}
\alias{cascade_likelihood}
\title{Likelihood Estimation for Cascade Food Web Model}
\usage{
cascade_likelihood(A)
}
\arguments{
\item{A}{The food web adjacency matrix.}
}
\value{
The likelihood that the given adjacency matrix was produced by the Cascade model.
}
\description{
Likelihood Estimation for Cascade Food Web Model
}
