% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{democracy}
\alias{democracy}
\title{Democracy}
\usage{
democracy(S, C)
}
\arguments{
\item{S}{Number of species in the community.}

\item{C}{Average path length used to get connectance.}
}
\value{
An adjacency matrix of a random network where any two species are connected with probability \code{C/S}. this model differs from \code{anarchy} by not allowing any trophic loops.
}
\description{
Democracy
}
\examples{
democracy(20, 1.86)
}
