% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance_matrix-class.R
\name{build_dominance_matrix}
\alias{build_dominance_matrix}
\alias{is.dominance_matrix}
\alias{print.dominance_matrix}
\title{Create a dominance_matrix object to use in a trophic projection}
\usage{
build_dominance_matrix(dominance = NULL, nsp = NULL, ...)

is.dominance_matrix(x)

\method{print}{dominance_matrix}(x, ...)
}
\arguments{
\item{dominance}{a matrix (or NULL object) identifying the relative dominance of a species over other species feeding off the same resource}

\item{nsp}{the number of species; required if dominance is NULL (in which case a flat dominance matrix is returned)}

\item{...}{further arguments passed to or from other methods}

\item{x}{an object to print or test as a dominance_matrix object}
}
\value{
An object of class \code{dominance_matrix}
}
\description{
A 'dominance matrix' object stores the relative dominance of a species over other species sharing the same resource.
It is a sub-component of a \link[trophic]{trophic_dynamics} object
}
\note{
Passing dominance = NULL will create a flat dominance matrix with all species assigned equal dominance
}
\examples{

library(trophic)

# Load a pre-compiled dominance_matrix object

dominance <- data(example_dominance_matrix)

# Construct the dominance_matrix object

test_dominance <- build_dominance_matrix(dominance = dominance)

# Test if object is of the type 'dominance_matrix'
  
is.dominance_matrix(x)

# Print information about the 'dominance_matrix' object

print(x)
}
