% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_web-class.R
\name{build_food_web}
\alias{build_food_web}
\alias{is.food_web}
\alias{print.food_web}
\alias{plot.food_web}
\title{Create a food_web object to use in a trophic projection}
\usage{
build_food_web(interaction_matrix, ...)

is.food_web(x)

\method{print}{food_web}(x, ...)

\method{plot}{food_web}(x, ...)
}
\arguments{
\item{interaction_matrix}{a matrix identifying (weighted) trophic links between species pairs (rows eat columns)}

\item{...}{further arguments passed to or from other methods}

\item{x}{an object to print or test as a food_web object}
}
\value{
An object of class \code{food_web}
}
\description{
A 'food web' object stores the trophic dynamics of a set of potentially interacting species.
It is a sub-component of a \link[trophic]{trophic_dynamics} object
}
\examples{

library(trophic)

# Construct the food_web object

test_fw <- build_food_web(interaction_matrix = food_web)

# Test if object is of the type 'food_web'
  
\dontrun{
is.food_web(x)
}

# Print information about the 'food_web' object

\dontrun{
print(x)
}

# Plot a 'food_web' object

\dontrun{
plot(x)
}
}
