% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{gen_cascade}
\alias{gen_cascade}
\title{Generalized Cascade Food Web Model}
\usage{
gen_cascade(S, C)
}
\arguments{
\item{S}{Number of species in the community.}

\item{C}{The connectance, or fraction of realized links in the food web.}
}
\value{
The adjacency matrix of a cascade model food web where species higher on the niche axis feed on those lower with a fixed probability \code{1/(2 * C) - 1}.
}
\description{
Generalized Cascade Food Web Model
}
\section{Reference}{

Stouffer, D. B., J. Camacho, R. Guimera, C. A. Ng, and L. A. N. Amaral. 2005. Quantitative patterns in the structure of model and empirical food webs. Ecology 86:1301–1311.
}

\examples{
gen_cascade(20, .1)
}
