% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{min_pot_niche}
\alias{min_pot_niche}
\title{Minimum Potential Niche Model Food Web}
\usage{
min_pot_niche(S, C, f)
}
\arguments{
\item{S}{Number of species in the community.}

\item{C}{The connectance, or fraction of realized links in the food web.}
}
\value{
An adjacency matrix for a minimum potential niche model food web.
}
\description{
Minimum Potential Niche Model Food Web
}
\section{Reference}{

Allesina, S., D. Alonso, and M. Pascual. 2008. A general model for food web structure. Science (New York, N.Y.) 320:658–61.
}

