% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.matrix.R
\name{n.matrix}
\alias{n.matrix}
\title{n.matrix: Input matrix for calculations with intervals of niche width and niche overlap}
\usage{
n.matrix(x, it)
}
\arguments{
\item{x}{A data matrix of prey abundance consumed by a species, where rows correspond to prey and columns correspond to individuals.}

\item{it}{Desired iterations number}
}
\value{
A matrix with columns = it, where each column corresponds to the random selection of an individual from the original matrix.
}
\description{
n.matrix: Input matrix for calculations with intervals of niche width and niche overlap
}
\examples{
sp1 <- matrix(sample(0:5, 100, replace = T), nrow = 10, ncol = 10) ## a Random abundance matrix
n.matrix(sp1, it = 100)
}
