% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.overlap.hill.R
\name{n.overlap.hill}
\alias{n.overlap.hill}
\title{Calculate the trophic niche overlap and its 95% confidence intervals for species pairs based on Hill series.}
\usage{
n.overlap.hill(sp1, sp2, it, order = c("0", "1", "2"))
}
\arguments{
\item{sp1}{First species input matrix created with n.matrix or a data matrix of prey abundance.}

\item{sp2}{Second species input matrix created with n.matrix or a data matrix of prey abundance.}

\item{it}{Desired iterations number, MUST be equal to number of columns.}

\item{order}{Corresponds to the desired diversity order for the overlap calculation.}
}
\value{
A matrix of a single row and four columns, where the first column corresponds to the mean overlap value, the second to the standard deviation (SD) and the third and fourth to the inferior and superior confidence intervals, respectively.
}
\description{
Calculate the trophic niche overlap and its 95% confidence intervals for species pairs based on Hill series.
}
\examples{
sp1 <- matrix(sample(0:5, 100, replace = T), nrow = 10, ncol = 10) ## a Random abundance matrix
sp2 <- matrix(sample(0:7, 100, replace = T), nrow = 10, ncol = 10) ## a Random abundance matrix
sp1.ent <- n.matrix(sp1, it = 100) ## Input matrix for Species 1
sp2.ent <- n.matrix(sp2, it = 100) ## Input matrix for Species 2
n.overlap.hill(sp1.ent, sp2.ent, it = 100, order = 0) ## overlap by prey richness
n.overlap.hill(sp1.ent, sp2.ent, it = 100, order = 1) ## overlap by equally common prey
n.overlap.hill(sp1.ent, sp2.ent, it = 100, order = 2) ## overlap by dominant prey
}
