% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.overlap.R
\name{n.overlap}
\alias{n.overlap}
\title{Calculate the trophic niche overlap and its 95% confidence intervals for species pairs.}
\usage{
n.overlap(sp1, sp2, it, method = c("pianka", "schoener", "morisita"))
}
\arguments{
\item{sp1}{Species 1 input matrix created with the n.matrix function or a data matrix of prey abundance.}

\item{sp2}{Species 2 input matrix created with the n.matrix function or a data matrix of prey abundance.}

\item{it}{Desired iterations number, MUST be equal to the number of columns of both species.}

\item{method}{Overlap metric used to measure the overlap between the pair of species.}
}
\value{
A matrix of a single row and four columns, where the first column corresponds to the mean overlap value, the second to the standard deviation (SD) and the third and fourth to the inferior and superior confidence intervals, respectively.
}
\description{
Calculate the trophic niche overlap and its 95% confidence intervals for species pairs.
}
\examples{
sp1 <- matrix(sample(0:5, 100, replace = T), nrow = 10, ncol = 10) ## a Random abundance matrix
sp2 <- matrix(sample(0:7, 100, replace = T), nrow = 10, ncol = 10) ## a Random abundance matrix
sp1.ent <- n.matrix(sp1, it = 100) ## Input matrix for species 1
sp2.ent <- n.matrix(sp2, it = 100) ## Input matrix for species 2
n.overlap(sp1.ent, sp2.ent, it = 100, method = "pianka") ## Pianka's overlap
n.overlap(sp1.ent, sp2.ent, it = 100, method = "schoener") ## Renkonen's = Schoener overlap
n.overlap(sp1.ent, sp2.ent, it = 100, method = "morisita") ## Morisita's overlap measure
}
