% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{nested_hierarchy}
\alias{nested_hierarchy}
\title{Nested Hierarchy Food Web Model}
\usage{
nested_hierarchy(S, C)
}
\arguments{
\item{S}{Number of species in the community.}

\item{C}{The connectance, or fraction of realized links in the food web.}
}
\value{
An adjacency matrix for a nested hierarchy food web
}
\description{
Nested Hierarchy Food Web Model
}
\section{Reference}{

Cattin, M., L. Bersier, C. Banašek-Richter, R. Baltensperger, and J.P. Gabriel. 2004. Phylogenetic constraints and adaptation explain food-web structure. Nature 427:835–839.
}

\examples{
nested_hierarchy(20, .1)
}
