% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{netvis_trophic}
\alias{netvis_trophic}
\title{Adjacency Matrix Visualization}
\usage{
netvis_trophic(a)
}
\arguments{
\item{a}{Adjacency matrix}
}
\value{
A plot of the adjacency matrix where black cells indicate interactions
}
\description{
Adjacency Matrix Visualization
}
\examples{
n <- niche(30, .2)
avis(n)
}
