% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_ratio-class.R
\name{build_pb_ratio}
\alias{build_pb_ratio}
\alias{is.pb_ratio}
\alias{print.pb_ratio}
\alias{plot.pb_ratio}
\title{Create a pb_ratio object to use in a trophic projection}
\usage{
build_pb_ratio(range, length = NULL, probs = NULL, type = "stochastic",
  ...)

is.pb_ratio(x)

\method{print}{pb_ratio}(x, ...)

\method{plot}{pb_ratio}(x, ...)
}
\arguments{
\item{range}{a range of possible pb values}

\item{length}{length of pb value sequence}

\item{probs}{relative probabilities of pb values in pb_range}

\item{type}{one of 'fixed', 'gradient', or 'stochastic' that determines how pb_values are treated}

\item{...}{further arguments passed to or from other methods}

\item{x}{an object to print or test as a pb_ratio object}
}
\value{
An object of class \code{pb_ratio}
}
\description{
A 'pb ratio' object stores the potentially stochastic ratio of production to standing biomass
}
\examples{

library(trophic)

# Construct the pb_ratio object

test_pb <- build_pb_ratio(range = c(0.25, 5.75), probs = c(5, 20, 10, 3, 1, 1, 1))

# Test if object is of the type 'pb_ratio'
  
is.pb_ratio(x)

# Print information about the 'pb_ratio' object

print(x)

# Plot a 'pb_ratio' object

plot(x)
}
