% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{probabilistic_niche}
\alias{probabilistic_niche}
\title{Probabilistic Niche Model Food Web}
\usage{
probabilistic_niche(S, C, a = 0.999)
}
\arguments{
\item{S}{Number of species in the community.}

\item{C}{The connectance, or fraction of realized links in the food web.}

\item{a}{The probability that i eats j, when j is exactly on i’s feeding optimum. Default is 0.999, the vaulue used in the original paper.}
}
\value{
An adjacency matrix for a probabilistic niche model food web.
}
\description{
Probabilistic Niche Model Food Web
}
\section{Reference}{

Williams, R. J., A. Anandanadesan, and D. Purves. 2010. The probabilistic niche model reveals the niche structure and role of body size in a complex food web. PLoS ONE 5.
}

\examples{
probabilistic_niche(20, .1, .99)
}
