% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic_dynamics-class.R
\name{build_trophic_dynamics}
\alias{build_trophic_dynamics}
\alias{is.trophic_dynamics}
\alias{print.trophic_dynamics}
\alias{plot.trophic_dynamics}
\title{Create a trophic_dynamics object to use in a trophic projection}
\usage{
build_trophic_dynamics(food_web, efficiency_matrix = NULL,
  dominance_matrix = NULL, ...)

is.trophic_dynamics(x)

\method{print}{trophic_dynamics}(x, ...)

\method{plot}{trophic_dynamics}(x, ...)
}
\arguments{
\item{food_web}{a \link[trophic]{food_web} object identifying (weighted) trophic links between species pairs}

\item{efficiency_matrix}{an \link[trophic]{efficiency_matrix} object storing the efficiency of energy conversion between all pairs of species}

\item{dominance_matrix}{a \link[trophic]{dominance_matrix} object identifying the relative dominance of a species over other species feeding off the same resource}

\item{...}{further arguments passed to or from other methods}

\item{x}{an object to print or test as a trophic_dynamics object}
}
\value{
An object of class \code{trophic_dynamics}
}
\description{
A 'trophic dynamics' object stores the trophic dynamics of a set of potentially interacting species.
}
\examples{

library(trophic)

# Load pre-compiled food_web, efficiency_matrix, and dominance_matrix objects

food_web <- data(example_food_web)
dominance <- data(example_dominance_matrix)
efficiency_mean <- data(example_efficiency_matrix)
efficiency_sd <- data(example_efficiency_sd)

# Construct the component objects
test_fw <- build_food_web(interaction_matrix = food_web)
test_efficiency_matrix <- build_efficiency_matrix(efficiency_mean = efficiency_mean, efficiency_sd = efficiency_sd)
test_dominance <- build_dominance_matrix(dominance = dominance)

# Construct the trophic_dynamics object
test_trophic_dynamics <- build_trophic_dynamics(food_web = test_fw, efficiency_matrix = test_efficiency_matrix, dominance_matrix = test_dominance)

# Test if object is of the type 'trophic_dynamics'
  
is.trophic_dynamics(x)

# Print information about the 'trophic_dynamics' object

print(x)

# Plot a 'trophic_dynamics' object

plot(x)
}
