% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological.R
\name{anarchy}
\alias{anarchy}
\title{Anarchy}
\usage{
anarchy(S, C)
}
\arguments{
\item{S}{Number of species in the community.}

\item{C}{Average path length used to get connectance.}
}
\value{
An adjacency matrix of a random network where any two species are connected with probability \code{C/S}.
}
\description{
Anarchy
}
\section{Reference}{

Cohen, J. E., and C. M. Newman. 1985. A stochastic theory of community food webs: I. Models and aggregated data. Proceedings of the Royal Society B 224:421–448.
}

\examples{
anarchy(20, 1.86)
}
