% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production_estimates-class.R
\name{estimate_production}
\alias{estimate_production}
\alias{is.production_estimates}
\alias{print.production_estimates}
\alias{plot.production_estimates}
\title{Estimate production for each node in a food web}
\usage{
estimate_production(trophic_dynamics, primary_producers,
  stochastic = c("food_web", "efficiency", "primary_production"),
  nsim = 100)

is.production_estimates(x)

\method{print}{production_estimates}(x, ...)

\method{plot}{production_estimates}(x, nodes = NULL, settings = list(), ...)
}
\arguments{
\item{trophic_dynamics}{a dynamics object - food web, trophic efficiency, trophic dominance, and P:B ratios}

\item{primary_producers}{a primary_producers object containing estimates of primary production}

\item{stochastic}{a character vector naming the stochastic elements of the trophic projection; can be any combination of "food_web", "efficiency", and "primary_production"}

\item{nsim}{number of estimates to generate}

\item{x}{a production_estimates object}

\item{...}{further arguments passed to or from other methods}

\item{nodes}{integer or integer vector of nodes to plot (indexed by food_web row)}

\item{settings}{plot settings passed directly to \link[graphics]{plot}}
}
\value{
An object of class \code{production_estimates}
}
\description{
Estimation updates replicate production values based on a trophic_dynamics object
}
\examples{

library(trophic)
library(future)
plan(multiprocess)

# Construct the component objects
test_fw <- build_food_web(interaction_matrix = food_web)
test_efficiency_matrix <- build_efficiency_matrix(efficiency_mean = efficiency_mean,
                                                  efficiency_sd = 0.01)
test_dominance <- build_dominance_matrix(dominance = dominance_matrix)
test_primary_producers <- build_primary_producers(production_mean = c(1, 2),
                                                  production_sd = c(0.5, 0.5))

# Construct the trophic_dynamics object
test_trophic_dynamics <- build_trophic_dynamics(food_web = test_fw,
                                                efficiency_matrix = test_efficiency_matrix,
                                                dominance_matrix = test_dominance)

# Estimate production values from constructed trophic_dynamics object
production_estimates <- estimate_production(test_trophic_dynamics, test_primary_producers)

# Test if object is of the type 'production_estimates'
\dontrun{
is.production_estimates(x)
}

# Print information about the 'production_estimates' object

print(x)

# Plot a 'production_estimates' object

\dontrun{
plot(x)
}
}
